requirements_all.txt
# pip install -r requirements.txt
playsound
pygame
joblib>=0.11
threadpoolctl>=2.0.0
joblib>=0.11
patsy>=0.5.1
scipy
scikit-image
keras==2.1.6
tensorflow==1.15.0
h5py==2.10.0
scikit-image==0.16.2
opencv-python
imgaug
IPython[all]
gTTs

# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# logging -------------------------------------
tensorboard>=1.5.0
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.9.0
# scikit-learn==0.19.2  # for coreml quantization
# tensorflow==2.4.1  # for TFLite export

# extras --------------------------------------
# Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
# albumentations>=1.0.3
thop  # FLOPs computation

# pyqt5
pyqt5
pyqt5-tools
